requirements.txt
django>=5.0,<6.0
pillow>=10.0.0
numpy>=1.25.0
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
opencv-python>=4.9.0.0
tensorflow>=2.15.0
importlib_resources

